% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{readJson}
\alias{readJson}
\title{A function that reads a JSON file and returns dataframe
seperated by their level. Form level orgaznied in one dataframe
and grid level is organized in another.}
\usage{
readJson(
  filePath,
  formLevelKey = NULL,
  gridLevelKey = NULL,
  idKey = "confirmationId",
  ignoreFormLevelData = FALSE
)
}
\arguments{
\item{filePath}{path of the JSON input}

\item{formLevelKey}{first level key}

\item{gridLevelKey}{second level key}

\item{idKey}{unique id to match second level to first level
generally unique id is attached to first level}

\item{ignoreFormLevelData}{if we don't want the first level data
provide this param as TRUE and function will only return a dataframe}
}
\value{
a dataframe
}
\description{
A function that reads a JSON file and returns dataframe
seperated by their level. Form level orgaznied in one dataframe
and grid level is organized in another.
}
