% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{readRedipReport}
\alias{readRedipReport}
\title{Title}
\usage{
readRedipReport(
  jsonFile,
  outputFile,
  idKey = "confirmationId",
  formLevelKey = "form",
  gridLevelKeys = c("deliverableTasks", "expenditureReport")
)
}
\arguments{
\item{jsonFile}{file path for json}

\item{outputFile}{file path for output excel.}

\item{idKey}{unique id name}

\item{formLevelKey}{form level key}

\item{gridLevelKeys}{data grid level keys a vector}
}
\value{
a list of dataframe and write to excel file.
}
\description{
Title
}
\examples{
readRedipReport(jsonfile = "data//path.json", outputFile = "data//path.xlsx")
}
